Finalize user API test data

+++ src/test/java/com/github/elenamikhailova/automation/api/tests/UserApiTest.java
@@ -1,11 +1,12 @@
 package com.github.elenamikhailova.automation.api.tests;
 
-import com.github.elenamikhailova.automation.api.client.ProductsApiClient;
 import com.github.elenamikhailova.automation.api.client.UserApiClient;
+import com.github.elenamikhailova.automation.api.data.UserData;
 import com.github.elenamikhailova.automation.api.model.CreateUserRequest;
 import com.github.elenamikhailova.automation.base.BaseApiTest;
 import io.restassured.response.Response;
 import net.datafaker.Faker;
+import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -15,7 +16,8 @@ import static org.hamcrest.Matchers.empty;
 
 public class UserApiTest extends BaseApiTest {
     private UserApiClient userApiClient;
-    private Faker faker = new Faker();
+    private CreateUserRequest user;
+    private UserData userData = new UserData();
 
 
     @BeforeEach
@@ -26,58 +28,44 @@ public class UserApiTest extends BaseApiTest {
     @Test
     @DisplayName("POST /createAccount creates a new user")
     void canCreateUser() {
-        CreateUserRequest user = generateRandomUser();
+        user = userData.generateRandomUser();
         Response response = userApiClient.createUser(user);
         response.then()
                 .statusCode(200)
                 .body("responseCode", equalTo(201))
                 .body("message", equalTo("User created!"));
-        userApiClient.deleteUserAccount(user);
     }
 
     @Test
     @DisplayName("GET /getUserDetailByEmail returns user details")
     void canGetUserByEmail() {
-        CreateUserRequest user = generateRandomUser();
+        user = userData.generateRandomUser();
         userApiClient.createUser(user);
         Response getUserResponse = userApiClient.getUserByEmail(user.getEmail());
         getUserResponse.then()
                 .statusCode(200)
+                .body("responseCode", equalTo(200))
                 .body("user.email", equalTo(user.getEmail()));
-        userApiClient.deleteUserAccount(user);
     }
 
     @Test
     @DisplayName("DELETE /deleteAccount deletes user")
     void canDeleteUser() {
-        CreateUserRequest user = generateRandomUser();
+        user = userData.generateRandomUser();
         userApiClient.createUser(user);
         Response deleteResponse = userApiClient.deleteUserAccount(user);
         deleteResponse.then()
                 .statusCode(200)
+                .body("responseCode", equalTo(200))
                 .body("message", equalTo("Account deleted!"));
+        user = null;
     }
 
-    private CreateUserRequest generateRandomUser() {
-        return CreateUserRequest.builder()
-                .name(faker.name().firstName())
-                .email("User_" + System.currentTimeMillis() + "@test.com")
-                .password(faker.credentials().password())
-                .title("Mrs")
-                .birthDate(String.valueOf(faker.number().numberBetween(1, 29)))
-                .birthMonth(String.valueOf(faker.number().numberBetween(1, 13)))
-                .birthYear(String.valueOf(faker.number().numberBetween(1980, 1999)))
-                .firstName(faker.name().firstName())
-                .lastName(faker.name().lastName())
-                .company(faker.company().name())
-                .address1(faker.address().streetAddress())
-                .address2(faker.address().secondaryAddress())
-                .country("Canada")
-                .zipCode(faker.address().zipCode())
-                .state(faker.address().state())
-                .city(faker.address().city())
-                .mobileNumber(faker.phoneNumber().cellPhone())
-                .build();
+    @AfterEach
+    public void tearDown() {
+        if (user != null) {
+            userApiClient.deleteUserAccount(user);
+        }
     }
 }
